Create plot_GOES_data.py
This program read GOES files (cloud mask or a specific channel) and produce an estimate of Cloud Fraction outside and inside an SST contour (the value set for is the 26.5 contour). A good part of this code has been readapted from personal communication with the developer of GOES toolbox (joaohenry23@gmail.com). To use this program the GOES package needs to be installed first. See the instruction at https://github.com/joaohenry23/GOES